+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Orphans.hs
@@ -1,0 +1,6 @@
+module Cardano.Ledger.Shelley.Orphans
+  {-# DEPRECATED "Use 'import Cardano.Ledger.Orphans' instead" #-}
+  ()
+where
+
+import Cardano.Ledger.Orphans ()